First installation points in Dispatcher_Task.cpp.
In order to use the DATASTREAM_ROOT environment variable must be set. I
chose to use this because as with the ACE_ROOT it allows use to put the
DSUI/DSKI code anywhere.

In order to collect data we will need a simple example which I will setup
to collect information.

Transforming collected data from binary format to will be a bit envolved but
nothing that cannot be scripted.


git-svn-id: svn://svn.dre.vanderbilt.edu/DOC/Middleware/branches/EC-DSUI-0@53110 ae88bc3d-4319-0410-8dbf-d08b4c9d3795